Create the array, constructor and addToLibrary()
The empty array, myLibrary would contain the objects
Initialize a constructor function with 4 parameters.

Create two instances of the Book object. The
addToLibrary() function checks if the object
is an instance of the Book object. If it is,
it adds it to the the myLibrary array and returns
the array. So technically, the myLibrary array
is an array of objects.