Refactor the addToLibrary() function.
Create two instances (book1 and book2) of
the Book object so that there would be some
books to be pushed into the myLibrary array.

The addToLibrary() function now takes user input
and creates another instance of the Book object
out of the input. It pushes this instance created
from the user input the two previous instances
that were created, into the myLibrary array and
returns the array.

On the HTML, create a form and 4 input fields.
Use the value attribute to populate the fields
to simulate user input of values which can be taken
by the addToLibrary() function.