Display a book-card div on the page.
First call the addToLibrary() function
so that the myLibrary array is returned.

On the HTML, create a container and a child
.books div. This .books div will be the
parent of the .book-cards divs.

On the javascript, create a for loop a function
display books. For now the function loops
through the myLibrary array and console logs
the title of every book. (The value of every
title key in every object in the array.)

The function also selects the .books div,
creates another div, gives the created div
a class of .book-card, and appends the created
div (as a child) to the .books div.

On the CSS, give the .book-card basic styles
so they can be displayed in the browser.