Display the books on the .book-cards divs
The displayBooks() function now loops through
the myLibrary array and for each object, it
creates a div and gives it a class of
.book-card. It then creates 4 divs. One for
the value of each key (title, author, pages and
read). For each of them, it sets the
textContent to be the respective value, gives
it, the respective class, and appends
it to the respective .book-card.
It appends the title, author, pages and read
in that order. The 4 divs are now children of
the .book-card div. It then appends the
.book-card to the .books container.

In the CSS, give the the 4 divs basic styles
to make them visible in the browser.
Display the .book with grid and
set the columns with a repeat function.
Display the book cards with flex and
set the flex-direction to be column
so that the card contents can stack
one ontop another.