Add a dialog element for the user input.
The dialog contains the form where the
user will input book details.

Add a confirmButton inside the dialog
that the user can click to confirm input.

Add a showDialog button that when users click
will show the modal.

In the CSS, style the input elements
with the display of block so that they
stack one on top each the other.