Refactor the addToLibrary() function
First redeclare the cardsContainer as
a variable outside the displayBooks() function
so that it can be accessed outside the function.

Relocate the the displayBooks() function above the
addToLibrary() function so that it can be called
by the addToLibrary() function.

Refactor the addToLibrary() function, so that
when the showButton button is clicked, it
opens the modal. It then makes the value of
the confirmButton to be the inputArray. This
was defined as an empty array at the beginning
of the function and all the user input will
be pushed into it.

The dialog then listens for the close() event
and when it happens, it pushes the user input
variables into the inputArray. (The user input
were changed from constants to variables because
they can change.) It then creates an object
instance of the Book object with the elements
of the array and stores it in the userBook
variable. It empties the inputArray so that
it can accept the next user input and then pushes
the userBook into the myLibrary array. It clears
the cardsContainer and calls the displayBooks()
function. If the display is not cleared and the
displayBooks() function is called, the  function
will loop through the myLibrary array and add all
books to the cards container. This adds the
current userBook, book1 and book2, to the
cardsContainer below book1 and book2 - creating
multiple entries in the DOM tree.

The confirmButton then listens for the click event
and when it happens, prevents the form from
submitting and calls the close() method on the
the dialog. It also passes the inputArray as a argument
to the close() method as the return value of dialog.

The input is array is passed into the close() method
so that the close method can use the array to create
an object instance of the Book object and store it
in userBook? It works without passing the array
as the argument to the close() method.
More information -

https://developer.mozilla.org/en-US/docs/Web/HTML/Element/dialog#examples

Push book1 and book2 into the myLibrary array on
line 14 so that these books are displayed on the page
when it loads - even before user input.