Add and implement the dialog cancel button
On the HTML, add the cancel button. On the JS
refactor the addToLibrary() function
so it can close the modal when the cancel button
is clicked.

Attach a click event listener to the cancel
button so that it closes the modal on click.

Since the return value of a dialog is a string,
(was previously trying to pass an array as the
return value and it was console logging
blank space), the function sets the return
value of the book dialog to "initialString".
It the sets the value of the confirm button to
"thereIsUserInput"

More information: https://developer.mozilla.org/en-US/docs/Web/API/HTMLDialogElement/returnValue

When the close event happens, if first checks if
the returnValue of the dialog is  "thereIsUserInput"
and if so, it works as in the previous commit.
Before it calls displayBooks(), it sets the
returnValue of the dialog back to
"initialValue". If the condition
is not met, it just returns.

Set it back to "initialValue" because
when the return value of the dialog
has been set to the value of the confirm button
("thereIsUserInput") on line 105 (means the dialog has
once pushed into the inputArray), other subsequent
clicking of the cancel button will close the modal
but will work with whatever values are on the form
(even if empty) and push into the myLibrary array.

On line 105 close the dialog and set the
return value of the dialog to the value
of the confirm button ("thereIsUserInput")
when the confirm button is clicked. Setting
it to "thereIsUserInput" so that the condition
set on line 73 is satisfied.