Add delete button to each card.
In the displayBooks() function, create a
delete button, give it a class and
append it to the bookCard. Thus every card
displayed will have a delete button. This
will be used to remove the book from the
library.

In the CSS, add a style declaration for
the delete button and give it a basic
style.