Give books a data-attribute
Edit the for-of loop in the displayFunction()
so that it uses the Array.prototype.entries()
with destructuring assignment to retrieve
the books and their indexes in the myLibrary
array.
https://medium.com/@francois.barrailla/javascript-iterate-over-array-values-and-indexes-using-a-for-of-loop-106a58972b24

After retrieving each book and its
index, it sets the data-attribute
(data-array-index) of the card containing
the book, to its index.