Make delete button to delete array index
In the displayBooks() function,
add the data-attribute to the delete button
and remove it from the card containing the
book. This is so that each delete button
will hold the array index of the book that
it is attached to (present in its card).

Store the bookIndex (as obtained from the
delete button's data-attribute)
in a variable in line 57 and Add a click
event listener to the delete button.
When the delete button of any book card is
clicked, the corresponding book is
spliced from the myLibrary array (at the index
stored in the bookIndex variable - the
index which corresponds to the data-attribute
of its delete button).