Update the display after the deleted book
In the displayBooks() function, after the
book is deleted, clear the cardsContainer
and call the displayBooks() function
recursively. This is so that the function
loops through the array again and displays
all the books remaining after the deleted
book(s).

Recursive functions should have a base
condition to prevent infinite recursion
but the displayBooks() cannot
perform infinite recursion. This is
because the recursive call is tied to
user input (click of the delete button).
When the delete button deletes
the book, the displayBooks() function
is recursively called and it rebuilds
the cardsContainer without the book
that was deleted. When the last book
is deleted, myLibrary is empty and the
displayBooks() recursion ends because
there is no book to display and no
delete button to click on.