Target the read/unread radio buttons
In the addToLibrary() Make a variable,
selectReadStatus that targets the value
of the input element with the
name - read-status that has a pseudo class
of checked. This is so whatever option that
the user selects is pushed into the inputArray
add used to create a new instance of the Book
object. This variable replaces the readField
variable.

Also change the read statuses of the
book1 and book2 to read and unread respectively
to be uniform with incoming read status input
from users.