Add the toggle read status functionality
On line 76 In the displayBook(), make the
rebuilding of the cards container and the
recursive call of the displayBook() into a
function -rebuildCardsContainer()

Edit the displayBooks(), such that on every
book card, it creates a toggle button.
The button's text content is dependent on the
read status of the book. Set a click event
listener on the toggle buttons so that on
click, the toggleRead() method which was
created on the Book object prototype is called.
This effectively changes the read status of
every book that its toggle button is clicked
on. The rebuildCardsContainer() is then called
to rebuild the cards container and display the
book with the new read status.