Clear the form before dialog opens.
Previously the last set of input values
that were used to create an instance of
of the Book object remained on the
form when the dialog opens again. This
is because the values of the input were
still stored in the title, author, pages
read status variables. New values would
just overwrite the old values and remain
on the form whenever the dialog opens.

In the HTML, remove the value attributes
and its values that were set the title,
author and pages inputs because when the
values are reset in the JS, they are reset
to their default values. Removing it sets
the default to an empty string.