Add form validation to dialog input fields
In the addToLibrary(), separate the
values of the form fields from the
form fields themselves. Store the
form fields in constants and their
values in variables. Move definitions
of these constants and variables
out from the bookDialogBox event
listener to the function scope
so that they can be accessed by
the confirmButton event listener.

On line 176, Edit the feature of the
confirmButton so that when it is
clicked on, it checks whether
these fields satisfy the constraint
validation rules set on them and if
so, it behaves as in the previous
commit, else it returns.

Give the Yes option of the read status
radio buttons a checked boolean attribute
so that there is not need to set
constraint validation rules on the
radio buttons.