Start styling the page
In the CSS, Add some resets to the HTML
elements. Also give the body element
a display of grid and define the first
and the last rows to be auto because they
will contain the header and footer. The
middle row is set to a height of 1fr so it
takes up the extra space as needed.

Edit the repeat function for the .books to
auto-fill instead of auto-fit so that they
more cards can still fill up the remaining
space with 1fr but when there is enough
space for another card, they shrink to their
minimum size. Change the minimum size to 300px.