Style the .header
In the CSS, give the .header a display
of flex and justify-content of
space between so that there can be space
between the h1 and the showDialog button.

Give the .header > h1 a font-family and
font-size.

Give the #showDialog an align-self of
center so that it is centered along the
cross axis and maintains its original
size - not stretching to the length of
its container. Also style the font-family,
font-size, padding, border and
border-radius of #showDIalog.

In the HTML, change the button text
to +Add book.