Remove the resets from the html selector
Remove the margin and padding resets
from the html selector and add them in
the body selector. As they were in the
html selector, they were not resetting
the margins and padding as needed - the
body selector is where they should be
added. Now they are resetting the margins
and padding as needed.