Add an all books paragraph and style
Add the paragraph in the HTML, as a
child of .main. Give it a display of
flex and take it to the center using
the justify-content and align-items
center property. Also style its
font-family, font size and margin.
The paragraph tells the user that
these are all the books.

The margin: 0 and padding: 0 resets
should be on the * (all) selector
but if it is there, it also resets the
margins and padding of the dialog.

Give body box sizing or border box. Give
header's h1 an appropriate margin to
reduce the space between the header and
main.