Style the book card
Center the contents of the book card
with justify-content and align-items
center. Give it an appropriate gap
and border-radius.

Style the title, author and pages
divs of the the book card with
appropriate font-families and
font-sizes.

Give the toggle button an align-self
of stretch so that it stretches to
fill the width of its container.
Style the padding top and bottom
to increase the top and bottom space
within the button.
Style the margin left and right to
give it some space from the edges
of its container.

In the javascript, update the display
of the pages div so that the
user-entered number of pages is
concatenated with the text "pages".