Fix padding on the .books and .book-card
In the CSS, change the .books left and right
padding. Reduce the min value in
the repeat function of the column tracks
definition to 250px so that there would be
space for 4 cards/row. The added left and right
padding reduced the space in the .books div and
with a min of 300px, there would have been
3 cards/row.

Reduce the book card margin and remove the
toggleButton's margins so that it stretches
to fill the width of its container as
constrained its container's padding.

In the javascript, update the textContent
of the readDiv. Style its font-family and
font-size in the CSS.