Populate the .footer with text and style
Wrap the footer text in a paragraph element
and the 'Github' in an anchor element (with
a href attribute).

If the footer text is not in a paragraph,
the space between the 'on' and the 'Github'
will not apply.

In the CSS, style the .footer with a display
of flex and center its content with the
justify-content property.

Style the font family and font size.