Style the dialog and form inputs
Style the dialog's padding and border
radius. Style the form with a display
of flex and give a flex direction of
column so that the contents of the form
can stack one on top another. Give the
form a gap to create space between the
form inputs.

Style the text and number inputs.

In the HTML, wrap the form buttons
in a .dialog-buttons div. Style the
div with a display of flex and a gap.
Give the buttons a flex of 1 so that
they can expand to fill their
container horizontally. Also give the
buttons a gap.