Add a paragraph in the dialog and style
In the HTML, add a paragraph at the top
of the dialog. In the CSS, style it
with a display of flex and center it
with the justify-content property. Style
its font family and font size.

Increase the font-size of the text and
number inputs.

Add padding-top to the .dialog-buttons
div so that there can be some more space
between them and the other form inputs.

Style the font-family and font-size of
the dialog buttons.