Style the fieldset and radio input
In the HTML, restructure the layout
of the radio buttons so that the
label and input are siblings so
that they can behave like flex items
when their parent is given a flex display.

Style the fieldset with a flex display
so that the label and input can behave
like flex items. Give fieldset a gap.
Style its font-family, font-size and
border radius.

Hide the native radio input using the
appearance, color and margin properties.
Create custom unchecked radio style.

Create a pseudo-element with ::before.
Add a transition to it and hide it
with the transform property.

Add a transform peoperty to the
checked state style so that when
the radios are checked they become
visible.
When they become visible, they do so
with an animation due to the previously
added transition.

More info - https://moderncss.dev/pure-css-custom-styled-radio-buttons/#theme-variable-and-box-sizing-reset