Add media query for smaller screens
At the bottom of normal styles, add
media query for screens smaller than
600px.

The media query targets the header
and changes the padding. it sets
only padding left so that there can
be space between the showDialog
button and the left edge of the
document.

Target the header h1 in the
media query and reduce the font-size.
A bigger font size would make the
h1 to wrap on small screens.

Target the showDialog button,
reduce the font size and padding
to make it smaller.

Also target the .main > p and make
the font a bit smaller.