Add more styling to the header and main
Give the header a box shadow with
vendor-prefixes. Give it a position
of relative so that the box shadow
can be visible above .main.

Style the showDialog button and its
hover state.

Style the main with a background color
that is lighter than than the header's
box shadow color for effect.

In the HTML, add the href link in the
footer's anchor tag.