Update the toggleButton class
In the javascript, update the toggleButton
in such a way that a class is added
depending on the read status
(book.read === "read").

In the CSS, create declarations for these
classes and give each a background color
(green for read, red for unread)

Increase the margin top and bottom on the
main > p and reduce the padding top and
bottom on .books to make the main > p
balance fairly between the header and the
book cards.