Style the dialog
Remove the border from dialog and give
it a matching background color with main.

Remove the outline and border from text
and number inputs and give them a border
radius

Remove the borders from the dialog
buttons and give each of them a
background color (confirm, green and
cancel, red)

Give the delete svg a red fill by inline
styling.