Wrap the radio input in a div and style
In the HTML, wrap the the fieldset, legend and radio input in a .radio-input div. Give it a white background, padding and border
radius.
The white background is to make it
look uniform with the number and text inputs.

Change the title of the document in the HTML

Style the hover states of the read-button, unread-button, confirmButton and closeButton.